pyproject.toml
[tool.poetry]
name = "project-1-bank"
version = "0.1.0"
description = ""
authors = ["Helen Chernotalova <chernotalova.e@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
python-dotenv = "^1.0.1"
requests = "^2.32.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
mypy = "^1.13.0"
black = "^24.10.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = ['venv', 'tests']

[tool.black]
# Максимальная длина строки
line-length = 119
# Файлы, которые не нужно форматировать
 exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | dist
  )/
  | foo.py

)
'''

[tool.isort]
# максимальная длина строки
line_length = 119
